Fix and improve CSS stylings based using the Classy theme. ISSUE-64
Mainly fixed how the Sponsor section is themed. Reduced the usage
of custom styling and Bootsrap grid overrides. Removed the old
styling that's not needed and not clear yet based on the mockup.
With this change on this section I'm proposing a more flat and
responsive approach on this section's empty state where the empty
boxes are using Bootstrap grids for either full width or not.

Updated the tab styling with minor changes to work well with
Bootstrap's tab styling.

Removed some unecessary centered text that's not also in the mockup
that surfaced.