Fix sponsor image grid display section. ISSUE-64
Updated styles to theme the Sponsors section.

Updated the page.html.twig file to remove the usage of
`.img-responsive` and wrapped the site's logo in it's own block.

Updated the Sponsors view template file to wrap the rows in a
`.row` block since we will be using Bootstrap's `.col-xx-xx` class.